Moved unit test to the existing class
